Updating IO routines to handle Okubo-Weiss and single precision

+++ NETCDF_IO/package_field.cpp
@@ -13,7 +13,7 @@ void package_field(
         MPI_Comm comm
         ) {
 
-    int index;
+    size_t index;
 
     // First, we need to compute the min and max values 
     //   to allow us to convert to signed shorts
@@ -24,7 +24,7 @@ void package_field(
     reduction(max : fmax_loc) reduction(min : fmin_loc)
     {
         #pragma omp for collapse(1) schedule(guided)
-        for (index = 0; index < (int) original.size(); index++) {
+        for (index = 0; index < original.size(); index++) {
             if ( mask->at(index) ) {
                 fmax_loc = std::max(fmax_loc, original.at(index));
                 fmin_loc = std::min(fmin_loc, original.at(index));
@@ -52,7 +52,7 @@ void package_field(
     private(index, local_double, local_int)
     {
         #pragma omp for collapse(1) schedule(guided)
-        for (index = 0; index < (int) original.size(); index++) {
+        for (index = 0; index < original.size(); index++) {
             if ( mask->at(index) ) {
 
                 // Scale original down to [-0.5,0.5] and store in local_double
